job recommendation create api done

+++ job/models.py
@@ -347,7 +347,7 @@ class JobRecommendation(P7Model):
     from pro.models import Professional
     professional = models.ForeignKey(Professional, on_delete=models.PROTECT)
     job = models.ForeignKey(Job, on_delete=models.PROTECT, db_column='job')
-    score = models.PositiveIntegerField()
+    score = models.DecimalField(max_digits=5, decimal_places=2)
     is_notified = models.BooleanField(default=False)
 
     class Meta:
